requirements.txt
asgiref==3.5.2
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
Django==3.2.14
django-cors-headers==3.13.0
django-pgconnection==1.0.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==5.0.0
lml==0.1.0
numpy
openpyxl==3.0.10
pandas
psycopg2==2.9.6
psycopg2-binary==2.9.3
pyexcel==0.7.0
pyexcel-io==0.6.6
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
typing-extensions==4.3.0
urllib3==1.26.10
xlrd==2.0.1
xlwt==1.3.0
zipp==3.9.0
